pyproject.toml
[tool.poetry]
name = "neuroglia"
version = "0.0.23"
description = "Port from .NET to Python of the Neuroglia Framework"
authors = ["Charles d'Avernas <charles.davernas@neuroglia.io>"]
license = "Apache"
readme = "README.md"
packages = [
    { include = "neuroglia", from = "src" },
]


[[tool.poetry.source]]
name = "gitlab"
url = "https://ccie-gitlab.ccie.cisco.com/api/v4/projects/703/packages/pypi"
priority = "supplemental"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.poetry.dependencies]
python = "^3.12"
annotated-types = "^0.6.0"
classy-fastapi = "^0.6.1"
esdbclient = "^1.0.17"
fastapi = "^0.109.2"
grpcio = "^1.60.1"
httpx = "^0.26.0"
multipledispatch = "^1.0.0"
pydantic-settings = "^2.2.0"
pymongo = "^4.6.1"
python-dotenv = "^1.0.1"
rx = "^3.2.0"
uvicorn = "^0.27.1"
typing-extensions = "^4.9.0"
pytest = "^8.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
pytest-asyncio = "^0.23.5"
mypy = "^1.8.0"
autopep8 = "^2.0.4"
coverage = "^7.4.1"
mypy-extensions = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 300
